pyproject.toml
[project]
name = "dsatutor"
version = "2.0.0"
description = "Adaptive DSA tutoring RL environment for OpenEnv"
authors = [
    { name = "Himangshu Yadav" }
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.0.0",
    "openenv-core>=0.2.0",
    "groq>=1.0.0",
    "google-genai>=0.3.0",
    "openai>=1.0.0",
    "python-multipart>=0.0.7",
    "requests>=2.31.0",
]
readme = "README.md"
requires-python = ">=3.11"

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
